q-a_ai/requirements.txt
langchain>=0.2.9,<0.3.0
langchain-community>=0.2.9,<0.3.0
langchain-core>=0.2.23,<0.3.0
langsmith>=0.1.75,<0.2.0
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=3.20.3
sentence-transformers==2.2.2
InstructorEmbedding